Pass seed to the huggingface trainer in train_hf_qa.py
The huggingface Trainer class resets the seed of numpy, torch, etc. to the one
given in the passed TrainingArguments object, which currently defaults to the
value given in the huggingface implementation (42).

In order to respect the seed given in the Omegaconf config value (cf.
`src/train.py`, L32), the config value cfg.seed is passed to the huggingface
TrainingArguments object, thus allowing to specify custom seed values to the
training runs.

NB: Since the current default seed provided in `src/conf/config.yaml` is
identical to that of the huggingface TrainingArguments, this commit will not
lead to changes relative to previous training runs.